CVSL-4184 Vary-Do-you-want-to-use-the-standard-HDC-curfew-hours-page
