+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/entity/HdcLicence.kt
@@ -179,8 +179,7 @@ class HdcLicence(
 ),
   HasElectronicMonitoringResponseProvider,
   HdcCase,
-  AlwaysHasCom,
-  HdcCurfewUpdatable {
+  AlwaysHasCom{
 
   fun copy(
     id: Long? = this.id,
