+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/entity/HdcVariationLicence.kt
@@ -171,8 +171,7 @@ class HdcVariationLicence(
 ),
   Variation,
   HdcCase,
-  AlwaysHasCom,
-  HdcCurfewUpdatable {
+  AlwaysHasCom{
 
   fun copy(
     id: Long? = this.id,
